tiny cal3d refactorings
git-svn-id: svn+ssh://svn.corp.imvu.com/home/svnrepos/trunk/stable/Source/cal3d@105745 07c76cb3-cb09-0410-85de-c24d39f1912e

+++ src/cal3d/vector.h
@@ -94,7 +94,7 @@ public:
         return (float)(x * x + y * y + z * z);
     }
     inline float length() const {
-        return (float)sqrt(lengthSquared());
+        return sqrtf(lengthSquared());
     }
     inline float normalize() {
         // calculate the length of the vector
